sunlight/requirements.txt
numpy==1.26.4
paho_mqtt==1.6.1
Requests==2.32.2
